ztree, datatables 更新，9-autocreate-modal改为本地script，修正bootstrapValidatorInit样式错误

+++ test-DataTables/server/ztree.ashx.cs
@@ -14,7 +14,7 @@ namespace test_DataTables.server
         public string Sex;
         public string Memo;
         public DateTime? Birth;
-        public int? Age;
+        public double? Age;
         public DateTime? DateStart;
         public DateTime? DateEnd;
         public string District;
